Merged AbstractVehicle and LocalSpace into a new class: Vehicle.
 - Working from Transforms and Rigidbody here, rather than internal representations to be mirrored.

 - Properties for most things there in stead of getter/setter methods.

Currently the Rigidbody implementation uses MovePosition and .transform.forward= rather than AddForce and AddTorque. Should be added.

Added ComboVehicle and ComboBehaviour classes. Drafts for now, the idea is to have the user specify weights on a selection of steering behaviours and facilitate working that out inside one single vehicle type.