Added basic obstacle avoidance to MovesToPoint vehicle
A few people had asked why that vehicle didn't avoid obstacles,
only neighbors.   Since each vehicle ends up using only the
functions it needs, and MovesToPoint was a simple test vehicle,
it didn't initially have any obstacle avoidance built in.  It
would be good to have it, however, as an example of how to balance
between the two avoidance types.

Documented MovesToPoint.Update to explain the logic of what's
going on behind the scenes.