Added simple CharacterMotorVehicle
This vehicle type is the inverse of AutonomousVehicle - instead of
moving the vehicle around, it updates its information from a separate
class (in this case CharacterMotor) so that other vehicles can keep
track of what it's doing.

NOTE: Because of either a bug or a limitation in MonoDevelop, as of
Unity 3 beta 3 the code will build within the Editor, but MD won't
find the CharacterMotor class at build time (even though it does
when editing).  I've filed it as 364011.