Bugfix Agent did not notice obstacle after sharp turns
If an agent was too close to an obstacle and had to execute a sharp
turn that put it in the obstacle's path, it didn't perceive the
obstacle as one because one of the intersections was behind the
agent's path starting point.  We should still avoid when we're
partially overlapping.