Moved radar collections to C5
Bit of an experiment.  The C5 collections are still better designed
than the .Net ones, even on 3.5.  For instance, List(T) has a Find
method, but IList(T) doesn't.  List(T).AsReadOnly returns an IList,
which means you can't do .Find on it. So if you'd like to have that,
you have to forget about the safety net of AsReadOnly, and end up
writing to the implementation.

C5 allows us to return a GuardedList that behaves consistent with
what you would expect from List, and on top of that the collection
gets us other features.