SteerForPathSimplified BUGFIX
There was an issue when the character's path began on its position
(ie, the character was already on the path) but the character's
speed was 0. In this case, the future path position estimated was
the same as the current position, which caused the steering behavior
to believe it had already arrived.