Cherry-picking agent-turning changes from 977db06d
There was a bug which caused the turning speed to be disregarded. I fixed
it on the experimental branch, and this commit cherry picks those changes
and adapts them to the development branch.

Notice that since those changes depended on the way that the experimental
branch handles velocity, they might not be as suitable for the old
approach.  Experiment with your agents first to see if this yields the
results you expect.

See properties:
- AutonomousVehicle.MinSpeedForTurning
- AutonomousVehicle.TurnTime

Note: A "TurnSpeed" might be more appropriate. TODO-TEST.

Cherrypicked:

Agent turning improvements

- Renamed SpeedFactorOnTurn to TurnSpeed, more accurate.
- The vehicle now gradually turns towards the desired velocity.
- Separated the application of the velocity from the vehicle's turning.
  This is a subtle change, but will likely make it easier to make
  tank-like vehicles.
- The minimum speed necessary for turning should be a per-vehicel value,
  not a global constant