Vehicle speed measuring improvements
A UnitySteer's vehicle has a velocity.  This is the force vector being
applied to the vehicle.  The previous implementation calculated the
vehicle's speed by using the vector's magnitude.

So far so good.  But suppose the vehicle hits a wall, or some other
behavior stops it from moving.  Now that Speed value I get from applied
Velocity is invalid, since the agent is not actually moving in space.
It's like measuring a car's speed by how fast it spins its wheels.

Added a Speedometer behavior, which is not UnitySteer-exclusive and can
be attached to any object, and expanded Vehicle. If a Vehicle detects
that its gameObject has a Speedometer, when its speed is requested it
will return the value measured by the Speedometer instead of just using
the Velocity's magnitude.