Optimization: Moved agent displacement from FixedUpdate to Update
On profiling with a 100 agents, it's cheaper to run the displacement on
Update than on FixedUpdate.  For what it's worth, the cost seems to be
about the same for kinematic and non-kinematic rigidbodies -
non-kinematics are cheaper to translate, but incur an added cost on the
Physics.simulate call.

Also replaced some leftover references of transform for the cached value,
and added profiling code.