DetectableObjects cache the transform on a public property
Previously we were just caching it on a protected _transform, but this
allows us to read it from the steering behaviors as well, without having
to cache it themselves.

I'm open for comments on how people feel about naming conventions? Should
this type of property be called Transform, CachedTransform or something
else?