EXPERIMENTAL OPTIMIZATION: Removing allocations on Radar
I've limited the number of objects a radar can detect in order to be able
to pre-allocate an array for them.  I'm also caching the detectable
objects or vehicles associated with a collider, if any, instead of
querying for them continually.

On a test with 100 boids running around, plus a few other agents, 50 of
which are considered every tick, the  FilterDetect method dropped from
taking about 1ms and allocating 32k to taking 0.2ms and doing no
allocations for the most part (after the collider's transforms have been
cached).

The current approach is experimental, and might not be ideal, since I
have a dictionary of transforms as a static value which will need to be
reset when a scene is loaded (oh, to have a OnSceneIsUnloaded method!),
but it'll do for testing.