Changed Radar.Vehicles,Obstacles to List<T>, optimized SteerForNeighbors
Using a list instead of a IEnumerable allows us to access the list by
index instead of using an enumerator, which in turn halves the number of
KB allocated per call and also reduces the total execution time on a test
with 100 boids.

Also checking if the GameObject has not been destroyed since we added it
to the list.