Improvement: No good reason for Vehicle.CanMove to be virtual
This change makes the code easier to grok, no longer has hidden
assignments.

Some vehicles clear their velocity when the CanMove property is set to
false, but this can be done just as well on the Update method - which
TickedVehicle is already doing.  It's not like an assignment is going
to break the performance bank.