MAJOR CHANGE: New handling of acceleration/deceleration smoothing
- New properties to control acceleration and deceleration rates for
  autonomous vehicles.
- Speed no longer has a setter.
- Renamed UpdateOrientationVelocity to the clearer UpdateDesiredVelocity.

See http://arges-systems.com/blog/2014/01/30/unitysteer-acceleration-smoothing-changes/

Considered adding a curve for the deceleration, but haven't decided yet
what would be some easily understandable units to express it in. My
current thinking if that if someone wants that sort of fine-grained
control over acceleration/deceleration, there is always the possibility
of subclassing AutonomousVehicle or creating a post-processing behavior.