Only DetectableObjects with colliders need to be unparented
The reason why they need to be unparented is because Radar will look for
a DetectableObject component only on the root, which means that if for any
reason we have objects without colliders on the scene, they can stay
wherever they are in the hierarchy.

As an aside, an ideal approach would be to just have each DetectableObject
report its collider to a list of known items on creation, which saves us
having to query for it, but that's an improvement for later.