BREAKING CHANGE: Removed scaled values from DetectableObject
OK, this is a big one.  It will affect how your vehicles and obstacles
are detected if they have a scale other than Vector3.one.

Starting on this commit, you *will* need to scale the radius for your
vehicles and objects manually if you change the object's scale. You will
also need to alter their current radius if the behavior is attached to an
object that you were scaling.

Put down the pitchforks and hear me out.

The scaled values seemed like a good idea when I first implemented them,
since the vehicle's radius would scale if you changed its visual scale.
Unfortunately, a side effect is that it introduced more properties that
you need to take into account when writing steering behaviors, which
could introduce errors and odd decisions, for instance:

- Do you use Radius or ScaledRadius?  Why?
- Do you really want the center scaled?
- Pop quiz! Is the SquaredRadius the square of the Radius or of the
  Scaled Radius?
- Pop quiz! What happens with the gizmo display in editor if the object
  was just deserialized, but the radius hasn't been touched? Will you be
  shown the scaled radius, the base radius, or neither?
- What if we rescale the transform without informing the vehicle?

It also has some minor quirks like not allowing easy multi-object
editing, and requiring several stub editors. Overall I've grown convinced
that the disadvantages of keeping them around outweigh the advantages.
Developers should be conscious of the changes in the object and when they
happen, and in this case, the best approach is having them perform the
change manually.

Since we're approaching a major release, it feels like this is the last
chance in quite a while to introduce a breaking change.